refactor: rename daemon to worker
